+++ packages/plugin-lava-player/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @discordx/plugin-lava-player
 
+## 2.2.0
+
+### Minor Changes
+
+- removed discordx plugin feature
+
 ## 2.1.3
 
 ### Patch Changes